simplify page token
