+++ server/src/main/java/GroceryFamily/GroceryMom/model/Page.java
@@ -1,0 +1,8 @@
+package GroceryFamily.GroceryMom.model;
+
+import java.util.List;
+
+public class Page<DATA> {
+    List<DATA> content;
+    DATA nextPageHead;
+}
