--- server/src/main/java/GroceryFamily/GroceryMom/model/PageTokenKeysExtractor.java
@@ -1,5 +1,0 @@
-package GroceryFamily.GroceryMom.model;
-
-public interface PageTokenKeysExtractor<DATA, KEYS> {
-    KEYS extract(DATA data);
-}
