+++ server/src/main/java/GroceryFamily/GroceryMom/service/mapper/ProductMapper.java
@@ -2,8 +2,6 @@ package GroceryFamily.GroceryMom.service.mapper;
 
 
 import GroceryFamily.GroceryMom.model.PageToken;
-import GroceryFamily.GroceryMom.model.PageTokenIO;
-import GroceryFamily.GroceryMom.model.PageTokenKeysExtractor;
 import GroceryFamily.GroceryMom.model.Product;
 
 import java.time.Instant;
